# perf: improve copy object time speed