--- a/lib/internal/fs/utils.js
+++ b/lib/internal/fs/utils.js
@@ -122,10 +122,7 @@ for (const name of Reflect.ownKeys(Dirent.prototype)) {
 }
 
 function copyObject(source) {
-  const target = {};
-  for (const key in source)
-    target[key] = source[key];
-  return target;
+  return { ...source };
 }
 
 function getDirents(path, [names, types], callback) {