@Himself65 2019-05-25
I've tried `{ ...source }`, and this `Spread syntax` cannot copy the unenumerable ​property

but I found `Object.assign(source)` is the best solution